+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.0
 
+- Added optional input `version_matcher`: a string holding a regex to match the target version number
 - Updated workflow files from `mangs/dependency-super-cache-action@v2` to `mangs/super-cache-action@v3`
 - Manually updated `marked` vendor dependency from version `4.0.12` to `4.0.16`
 - Update dependencies